% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{check_ids}
\alias{check_ids}
\title{Check for and omit non-animal animal ids}
\usage{
check_ids(r, ids, omit = c("wand", "error"), id_length = 10, bird_ids)
}
\arguments{
\item{r}{Data frame. A data frame of read data, with at least column
\code{animal_id}.}

\item{ids}{Character or Data frame. Either the name and location of a
.csv file containing the animal id index or a data frame of the animal id
index. In either case, the data must contain two columns: \code{animal_id}
and \code{species}.}

\item{omit}{Character vector. All the values of \code{species} in the
animal id index which you would like to omit from your data. Defaults to
\code{c("wand", "error")}.}

\item{id_length}{Numeric. How many characters are expected in each id? This
will test to make sure all ids are the right length (i.e. make sure
leading zeros haven't been omitted). NA skips this test.}

\item{bird_ids}{Deprecated.}
}
\value{
A data frame without the specified animal ids. Messages are printed to
inform the user of matching or non-matching animal ids.
}
\description{
This function pads all short ids with leading zeros, then compares the animal
ids in your data file to those in an index file. Any ids matched to a species
with a value in the \code{omit_animal} argument will be omitted from the data
set.
}
\details{
This is useful for removing ids you know are errors (e.g., \code{0000000000})
or which you know are wands.
}
\examples{
\dontrun{
r <- check_ids(finches, ids = "animal_index.csv")

index <- read.csv("animal_index.csv")
r <- check_ids(finches, ids = index)
}
}
