% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_activity}
\alias{convert_activity}
\title{Convert data for use by the activity package}
\usage{
convert_activity(r)
}
\arguments{
\item{r}{Dataframe. Raw RFID dataset. Must have at least columns
\code{animal_id}, \code{logger_id} and \code{time}. The column time should
be formated as POSIXct (data/time). (Consider using
\code{\link[=load_format]{load_format()}} to format your dataframe)}
}
\value{
A list of vectors corresponding to each individual. See examples for
specific application.
}
\description{
Converts raw RFID data into a format for easy use by the
\code{\link[activity:fitact]{activity::fitact()}} function included in the \link{activity}
package for modelling activity levels and daily activity patterns.
}
\examples{

# Format for use by activity package
i <- convert_activity(finches_lg)

\dontrun{
# Use activity package:
library(activity)

# Calculate daily activity pattern for a single individual
a <- fitact(i[[1]], sample = "none")
plot(a)

# Calculate daily activity pattern for all individuals
a <- lapply(i, fitact, sample = "none")
plot(a[[3]])
plot(a[["06200004F8"]])
# etc.
}

}
\seealso{
activity package and it's function \link[activity:fitact]{fitact}.
\url{https://cran.r-project.org/package=activity}
}
