% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_anidom}
\alias{convert_anidom}
\title{Convert displacements for use by functions from the aniDom package}
\usage{
convert_anidom(d)
}
\arguments{
\item{d}{Data frame or List. Either the specific displacements data frame which is
returned as a list item from \code{disp()}, or the whole displacements list
returned by \code{disp()}.}
}
\value{
A data frame listing winners and losers of all displacements sorted
  by time of the event.
}
\description{
Converts displacements data into a format for easy use by either the
\code{\link[aniDom]{elo_scores}}, the
\code{\link[aniDom]{estimate_uncertainty_by_repeatability}}, or the
\code{\link[aniDom]{estimate_uncertainty_by_splitting}} functions included in
the \link{aniDom} package.
}
\examples{

# Calculate displacements
d <- disp(visits(finches_lg))

# Format for use in aniDom's elo_scores
i <- convert_anidom(d)
i <- convert_anidom(d$displacements) # Equivalent

# Use aniDom functions:

\dontrun{
library(aniDom)

# Calculate elo_scores
s <- elo_scores(winners = i$winner, losers = i$loser)

# Estimate repeatability
r1 <- estimate_uncertainty_by_repeatability(winners = i$winner, losers = i$loser)
r2 <- estimate_uncertainty_by_splitting(winners = i$winner, losers = i$loser, randomise = TRUE)

}

}
\seealso{
\link{aniDom} package and it's functions: \link[aniDom]{elo_scores}, \link[aniDom]{estimate_uncertainty_by_repeatability}, and \link[aniDom]{estimate_uncertainty_by_splitting}. \url{https://cran.r-project.org/package=aniDom}
}
