% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_dominance}
\alias{convert_dominance}
\title{DEFUNCT - Convert displacements for use by functions from the Dominance package}
\usage{
convert_dominance(d)
}
\arguments{
\item{d}{Data frame or List. Either the specific displacements data frame which is
returned as a list item from \code{disp()}, or the whole displacements list
returned by \code{disp()}.}
}
\value{
List of data frames to use in functions from the Dominance
package. data_sheet contains all interactions: action.from/action.to
represent individuals (code is matched to animal_id in the items data
frame). action.from represent displacers (winners), action.to represent
displacees (losers), kind.of.action is a dummy variable representing the
action type (displacement). items is a data frame matching animal_ids to to
the code used in the data_sheet. actions contain the one action type
(displacements) and the classification and weighting (corresponds to
action.from being a winner, and action.to being a loser). bytes is a dummy
vector indicating that the action type "displacement" should be included in
the calculation. See examples for specific application.
}
\description{
\strong{The Dominance package is no longer available therefore this function
is defunct}
}
\details{
Converts displacements RFID data into a format for easy use by either the
\code{Dominance::ADI()} or the \code{Dominance::Sociogram()}
functions included in the Dominance package for calculating average
dominance indices and drawing sociograms. Only includes individuals that
participated in at least one interaction.
}
