% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_perc}
\alias{convert_perc}
\title{Convert displacements for use by the Perc package}
\usage{
convert_perc(d)
}
\arguments{
\item{d}{Data frame or List. Either the specific interactions data frame which is
returned as a list item from \code{disp()}, or the whole displacements list
returned by \code{disp()}.}
}
\value{
A data frame of interactions for input into
\code{\link[Perc:as.conflictmat]{Perc::as.conflictmat()}}. See examples for specific application.
}
\description{
Converts displacements RFID data into a format for easy use by the
\code{\link[Perc:as.conflictmat]{Perc::as.conflictmat()}} function included in the \link{Perc}
package. Can then be applied to internal Perc functions for calculating
dominance from perculation and conductance. Only includes individuals that
participated in at least one interaction.
}
\examples{
# Calculate displacements
d <- disp(visits(finches_lg))

# Format for use by Perc package
i <- convert_perc(d)
i <- convert_perc(d$interactions) # Equivalent

\dontrun{
# Use Perc package:
library(Perc)

# Calculate ranks (adapted from Perc examples)
conflict_mat <- as.conflictmat(i, weighted = TRUE)
perm <- conductance(conflict_mat, 2)
simRankOrder(perm$p.hat, num = 10, kmax = 1000)

}

}
\seealso{
\link{Perc} package and it's function \link[Perc:as.conflictmat]{as.conflictmat}.
\url{https://cran.r-project.org/package=Perc}
}
