% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{daily}
\alias{daily}
\title{Create daily activity data}
\usage{
daily(a1, pass = TRUE)
}
\arguments{
\item{a1}{Data frame. Data from output of \code{activity()}.}

\item{pass}{Logical. Pass 'extra' columns through the function and append them to the output.}
}
\description{
Summarizes and averages activity data over a 24-hr period, generating a 24-hr
daily activity pattern for plotting. The resulting data set contains four
columns reflecting the proportions of time blocks scored as active, inactive,
unknonw, or total.
}
\details{
Output timezone and dates are irrelevant, as the data is tied to times, not
dates. Therefore timezone is UTC and the dates are on 1970-01-01. When
plotting, omit the date part of the label to accurately portray time only.

Resolution of the data is automatically detected as the same as that
specified in \code{activity()}.
}

