% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{dom}
\alias{dom}
\title{Dominance}
\usage{
dom(d, tries = 50, omit_cutoff = 3)
}
\arguments{
\item{d}{Data frame or List. Either the interactions data frame which is
returned as a list item from \code{disp()}, or the whole displacements list
returned by \code{disp()}.}

\item{tries}{Numeric. The maximum number of iterations to find the 'best guess'}

\item{omit_cutoff}{Numeric. Minimum number of interactions (sum of wins and
losses) individuals must have (omitted otherwise).}
}
\value{
A list with the following named items:
\enumerate{
  \item \code{dominance}: A best guess at the dominance hierarchy (most to
least dominant) (one vector per 'best guess')

  \item \code{reversals}: Which individuals show reversals (and with whom)?
  (one data frame per 'best guess')

  \item \code{interactions}: A matrix of dominance interactions. Displacers
  across the top, displacees down the side. Values are the numbers of wins
  (upper triangle) or losses (lower triangle) against the opposing
  individual. (one matrix per 'best guess')

  }
}
\description{
Takes output from \code{disp()} and calculates linear dominance hierarchies. Should
be considered a starting point only.
}
\details{
These are linear heirarchies based on those commonly calculated in chickadee dominance studies (see References). They are linear and based on the number of wins vs. losses for each individual. The algorithm starts by ranking individuals from the highest proportion of wins to the lowest proportion. Then it checks for reversals (when an higher individual actually lost more against a lower ranking individual). Individuals are switched in and out until ideally, a dominance hierarchy is establed in which each individual won more interactions against all individuals ranked lower, and lost more interactions against all individuals ranked higher. The algorithm is not exhaustive, so may not find the heirarchy and should be thus considered a good starting point.

See also the \code{\link[Perc]{as.conflictmat}} function of the \link{Perc}
package to transform \code{disp} data into a conflict matrix, ready for
analysis of dominance using percolation and conductance. \code{disp} data can
also be converted using the \code{\link{convert_anidom}} function to a data
frame for use by the \link{aniDom} package's \link[aniDom]{elo_scores}
function.
}
\examples{

 # Look at dominance for chickadees in experiment 2
 v <- visits(chickadees[chickadees$experiment == "exp2",])
 d <- disp(v)
 dm <- dom(d$interactions)

 # But not necessary to specify interactions:
 dm <- dom(d)

 # Calculate across different experiments (expect warnings about unequal factor levels):
library(dplyr)

v <- chickadees \%>\%
  group_by(experiment) \%>\%
  do(visits(.))

d <- v \%>\%
  group_by(experiment) \%>\%
  do(data = disp(.))

dm <- d \%>\%
  group_by(experiment) \%>\%
  do(data = dom(.$data[[1]]))

# Look at the dominance data stored in the 2nd experiment:
dm$data[dm$experiment == "exp2"][[1]] #or
dm[["data"]][[1]] #or
dm$data[[1]]

# Look at the dominance matrices from 3nd experiment:
dm$data[dm$experiment == "exp3"][[1]]$matrices #or
dm[["data"]][[2]]$matrices #or
dm$data[[2]]$matrices


}
\references{
Otter K, McGregor PK, Terry AMR, Burford FRL, Peake TM, Dabelsteen T. 1999. Do female great tits (Parus major) assess males by eavesdropping? A field study using interactive song playback. Proceedings of the Royal Society of London. Series B: Biological Sciences 266:1305–1309.

Smith SM. 1976. Ecological aspects of dominance hierarchies in Black-capped Chickadees. The Auk 93:95–107.
}
