% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{feeding.layer}
\alias{feeding.layer}
\title{Add feeding layer to leaflet map}
\usage{
feeding.layer(map, f, f.scale = 1, f.title = "Feeding time",
  f.pal = c("yellow", "red"), controls = TRUE)
}
\description{
Designed for advanced use (see map.leaflet() for general mapping)
}

