% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{feeding}
\alias{feeding}
\title{'Visits' to 'feeding bouts' for a single bird_id}
\usage{
feeding(v1, bw = 15)
}
\arguments{
\item{v1}{Dataframe. A visits data frame containing only \strong{one} unique
bird id. From the output of \code{visits}. Must contain columns
\code{bird_id}, \code{feeder_id}, \code{start}, and \code{end}.}

\item{bw}{Numeric. The minimum number of minutes between visits for two
successive visits to be considered separate feeding bouts. When \code{bw} =
NULL only visits to another feeder are scored as a separate feeding bout.}
}
\value{
A data frame of feeding bouts. A feeding bout is defined as the bout
  of time spent making regular visits to a feeder, in which each visit
  occurrred within some cutoff time of the last. This data frame has the
  following columns:
  \itemize{
    \item ID of the bird (\code{bird_id})
    \item ID of the feeder(\code{feeder_id})
    \item Time of the start of the feeding bout (\code{feed.start})
    \item Time of the end of the feeding bout (\code{feed.end})
  }
}
\description{
For a single \code{bird_id}, turns visits at mulitple feeders into feeding
bouts. Feeding bouts are separated by switching feeders (when \code{bw =
NULL}). If \code{bw} is not \code{NULL}, they are separated by switching
feeders or by \code{bw} minutes.
}
\examples{
 \dontrun{

 v <- visits(r)

 # One bird at a time:
 f <- feed(v[v$bird_id == "062000039D",])

 # Split a data frame by \\code{bird_id} with \\code{plyr}, then apply \\code{move}.
 library(plyr)
 f <- ddply(v, c("bird_id"), feed)

 # Summarize a movement dataframe
 library(plyr)
 f.totals <- ddply(m, c("bird_id", "feeder_id"), summarise,
                   feed_length = sum(feed_length))
 }
}

