% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedr.R
\docType{package}
\name{feedr-package}
\alias{feedr-package}
\alias{feedr}
\alias{animalnexus-package}
\title{Transforming Raw RFID Data}
\description{
Load, clean, transform, and visualize raw RFID data for looking at a variety
of biological questions.
}
\details{
In-depth turorials available online:
\url{https://animalnexus.github.io/feedr/}
}
\references{
animalnexus \url{http://animalnexus.ca}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/animalnexus/feedr}
}

}
\author{
\strong{Maintainer}: Steffi LaZerte \email{steffi@steffi.ca}

}
