% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finches}
\alias{finches}
\title{Finches visits to RFID-enabled feeders}
\format{A data frame with 412 rows and 5 variables:
\describe{
  \item{animal_id}{Unique RFID tag attached to a finch}
  \item{time}{Time of read}
  \item{logger_id}{Unique id of logger where read was made}
  \item{lat}{Latitude}
  \item{lon}{Longitude}
  \item{species}{Individual species}}}
\usage{
finches
}
\description{
A dataset containing raw reads of finches visiting feeders (RFID loggers) on
Thompson Rivers University campus in Kamloops, BC, Canada
}
\keyword{datasets}

