% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finches_lg}
\alias{finches_lg}
\title{Large dataset of finche visits to RFID-enabled feeders}
\format{A data frame with 42708 rows and 7 variables:
\describe{
  \item{animal_id}{Unique RFID tag attached to a finch}
  \item{time}{Time of read}
  \item{logger_id}{Unique id of logger where read was made}
  \item{species}{Individual species}
  \item{sex}{Individual sex, if known}
  \item{lat}{Latitude}
  \item{lon}{Longitude}}}
\usage{
finches_lg
}
\description{
A large dataset containing raw reads of finches visiting feeders (RFID
loggers) on Thompson Rivers University campus in Kamloops, BC, Canada
}
\keyword{datasets}
