% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get.data}
\alias{get.data}
\title{Depreciated: Download data from original BirdMoves website.}
\usage{
get.data(start = NULL, end = NULL,
  url = "http://gaia.tru.ca/birdMOVES/rscripts/rawvisits.csv",
  feeder_details = c("loc"), bird_details = c("species"),
  tz_disp = "America/Vancouver", sites = "Kamloops")
}
\description{
Depreciated use \link{\code{dl_data}} instead.
}

