% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.web.legacy}
\alias{load.web.legacy}
\title{Load read data from the web (legacy)}
\usage{
load.web.legacy(r_file, tz = "America/Vancouver", sep = ",")
}
\arguments{
\item{r_file}{Character. The location of a single file to load.}

\item{tz}{Character. The time zone the date/times should be converted to (should match one of the zones produced by \code{OlsonNames())}.}

\item{sep}{Character. An override for the separator in the \code{read.table()} call (see \code{sep =} under \code{?read.table} for more details).}
}
\description{
Loads raw read data from csv files downloaded from the __old__ version of
\url{http://gaia.tru.ca/birdMOVES/datadownload.html} website and formats them
for use with the feedr functions. This is merely a wrapper function that does
many things that you can do yourself. It's utility depends on how
standardized your data is, and whether you have extra details you need to
address.
}
\examples{
\dontrun{r <- load.web.legacy("downloaded_file.csv")}
}

