% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.web}
\alias{load.web}
\title{Depreciated: Load read data from the web}
\usage{
load.web(r_file, tz = "America/Vancouver", tz_disp = NULL, sep = ",")
}
\description{
Depreciated use \link{\code{load_web}} instead.
}

