% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map.leaflet.base}
\alias{map.leaflet.base}
\title{Base map for leaflet}
\usage{
map.leaflet.base(locs, marker = "feeder_id", name = "Readers",
  controls = TRUE)
}
\description{
Designed for advanced use (see map.leaflet() for general mapping)
}

