% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_animate.R
\name{map_animate}
\alias{map_animate}
\title{Animated map with leaflet}
\usage{
map_animate(v)
}
\arguments{
\item{v}{Data frame. Data frame. Visits data frame created with the
\code{visits()} function.}
}
\description{
Interactive shiny app to select and animate feeder use of multiple
individuals over time. Also available through \code{animalnexus()}.
}
\examples{
\dontrun{
map_animate(visits(finches))
}

}

