% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_animate_indiv.R
\name{map_animate_indiv}
\alias{map_animate_indiv}
\title{Animated map for individuals with leaflet}
\usage{
map_animate_indiv(v)
}
\arguments{
\item{v}{Data frame. Data frame. Visits data frame created with the
\code{visits()} function.}
}
\description{
Interactive shiny app to select and animate movements and feeder use of
different individuals over time. Also available through \code{animalnexus()}.
}
\examples{

\dontrun{
map_animate_indiv(visits(finches))
}

}

