% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map_ggmap}
\alias{map_ggmap}
\title{Map data using ggmap}
\usage{
map_ggmap(p = NULL, m = NULL, locs = NULL, p_scale = 1, m_scale = 1,
  p_title = "Time", m_title = "Path use", p_pal = c("yellow", "red"),
  m_pal = c("yellow", "red"), maptype = "satellite", mapsource = "google",
  zoom = 17, which = NULL, u.scale, p.scale, u.title, p.title, u.pal, p.pal)
}
\arguments{
\item{p}{Dataframe. Summarized presence data with columns \code{logger_id} and
\code{amount}. It can also contain \code{animal_id} for individual-based
data, and lat/lon instead of a locs argument.}

\item{m}{Dataframe. Summarized movement data with columns \code{logger_id},
\code{move_path}, and \code{path_use}. It can also contain \code{animal_id}
for individual-based data, and lat/lon instead of a locs argument.}

\item{locs}{Dataframe. Lat and lon for each logger_id, required if lat and
lon not in either p or m.}

\item{p_scale, m_scale}{Numerical. Relative scaling constants to increase (> 1) or
decrease (< 1) the relative size of presence (p) and path (m) data.}

\item{p_title, m_title}{Character. Titles for the legends of presence (p) and
path (m) data.}

\item{p_pal, m_pal}{Character vectors. Colours used to construct gradients for
presence (p) and movement (m) data.}

\item{maptype}{Character. The type of map to download. See \code{maptype}
under \code{\link[ggmap]{get_map}} for more options.}

\item{mapsource}{Character. The source of the map to download. See
\code{source} under \code{\link[ggmap]{get_map}} for more options.}

\item{zoom}{Numeric. The zoom level of the map to download. See \code{zoom}
under \code{\link[ggmap]{get_map}} for more options.}

\item{which}{Character vector. A vector of animal ids specifying which to show.
Only applies when using individual data.}

\item{u.scale, p.scale, u.title, p.title, u.pal, p.pal}{Depreciated.}
}
\value{
An interactive leaflet map with layers for presence, movements and
  logger positions.
}
\description{
Visualize presence at and movements between loggers using ggmap in R. This produces a
static map. The user must summarize presence and movements in the manner
that they wish to visualize it. This function can take invidiual animal data as
well as grand summarise (see Details and Examples).
}
\examples{

v <- visits(finches)
p <- presence(v, bw = 15)
m <- move(v)

# Summarise data for visualization (use totals):
library(dplyr)
p_all <- p \%>\%
  group_by(logger_id, lat, lon) \%>\%
  summarise(amount = sum(length) / animal_n[1])

m_all <- m \%>\%
  group_by(logger_id, move_path, lat, lon) \%>\%
  summarise(path_use = length(move_path) / animal_n[1])

# Look at total summary maps
map_ggmap(p = p_all, m = m_all)

# Summarise data for visualization (use individuals):
p_indiv <- p \%>\%
  group_by(animal_id, logger_id, lat, lon) \%>\%
  summarise(amount = sum(length))

m_indiv <- m \%>\%
  group_by(animal_id, logger_id, move_path, lat, lon) \%>\%
  summarise(path_use = length(move_path))

# Look at individual summary maps (note that Leaflet just stacks individuals
# one on top of the other)
map_ggmap(p = p_indiv, m = m_indiv)

}

