% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map_leaflet_base}
\alias{map_leaflet_base}
\title{Base map for leaflet}
\usage{
map_leaflet_base(locs, marker = "feeder_id", controls = TRUE,
  minZoom = NULL, maxZoom = 18)
}
\description{
Designed for advanced use (see map_leaflet() for general mapping)
}

