% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_summary.R
\name{map_summary}
\alias{map_summary}
\title{Create summary map with leaflet}
\usage{
map_summary(v)
}
\arguments{
\item{v}{Data frame. Visits data frame created with the \code{visits()}
function.}
}
\description{
A shiny wrapper for \code{map_leaflet()}. Allows the user to dynamically
select the time range that should be summarized. For more options, see
\code{map_leaflet()}. Also available through \code{animalnexus()}.
}
\examples{

\dontrun{
map_summary(visits(finches))
}

}

