% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{move.layer}
\alias{move.layer}
\title{Add movement layer to leaflet map}
\usage{
move.layer(map, m, m.scale = 1, m.title = "Path use", m.pal = c("yellow",
  "red"), controls = TRUE)
}
\description{
Designed for advanced use (see map.leaflet() for general mapping)
}

