% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{path.layer}
\alias{path.layer}
\title{Add movement layer to leaflet map}
\usage{
path.layer(map, p, p.scale = 1, p.title = "Path use", p.pal = c("yellow",
  "red"), controls = TRUE)
}
\description{
Designed for advanced use (see map.leaflet() for general mapping)
}

