% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{presence}
\alias{presence}
\title{Presence at a logger}
\usage{
presence(v, bw = 15, pass = TRUE)
}
\arguments{
\item{v}{Dataframe. A visits data frame from the output of \code{\link{visits}} (may
contain more than one animal_id). Must contain columns \code{animal_id},
\code{logger_id}, \code{start}, and \code{end}.}

\item{bw}{Numeric. The maximum number of minutes between visits for them to
be considered the same event. When \code{bw} = NULL only
visits to another logger are scored as a separate event.}

\item{pass}{Logical. Pass 'extra' columns through the function and append
them to the output.}
}
\value{
A data frame of presence events. This data frame has the following
  columns:
  \itemize{
  \item ID of the animal (\code{animal_id})
  \item ID of the logger(\code{logger_id})
  \item Time of the start of the event (\code{start})
  \item Time of the end of the event (\code{end})
  }
}
\description{
Turns multiple visits at specific loggers into overall presence events.
Presence is different from a visit in that a visit is considered a specific
period of time in which the individual was in range of a logger to the
exclusion of others. In contrast, presence reflects a period of time where
the individual was making regular visits to the logger but not necessarily in
range at all times, or to the exclusion of other individuals. Presence can be
considered a less precise 'smoothing' of the data.
}
\details{
The start and end of a period of presence is determined by either switching
loggers (when \code{bw = NULL}) or by both switching loggers and by a cutoff
time of \code{bw} minutes.
}
\examples{

v <- visits(finches)
p <- presence(v)

# Summarize a movement dataframe (get total time present per logger per animal)
library(dplyr)
p_totals <- p \%>\%
    group_by(animal_id, logger_id) \%>\%
    summarize(length = sum(length))

# Calculate across different experiments (expect warnings about unequal factor levels):
library(dplyr)

v <- chickadees \%>\%
  group_by(experiment) \%>\%
  do(visits(.))

p <- v \%>\%
  group_by(experiment) \%>\%
  do(presence(.))


}

