% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{presence_layer}
\alias{presence_layer}
\title{Add presence layer to leaflet map}
\usage{
presence_layer(map, p, p_scale = 1, p_title = "Presence",
  p_pal = c("yellow", "red"), controls = TRUE, u.scale, u.title, u.pal)
}
\description{
Designed for advanced use (see map_leaflet() for general mapping)
}

