% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{sun}
\alias{sun}
\title{Get sunrise/sunset times}
\usage{
sun(loc, date, tz)
}
\arguments{
\item{loc}{Vector/Data frame. Longitude and Latitude coordinates for location
of sun rise/set}

\item{date}{Vector. Date(s) to cacluate sun rise/set for.}

\item{tz}{Timezone of the dates.}
}
\description{
Calculate times of sunrise and sunset depending on the location and the date.
}
