% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_import.R
\name{ui_import}
\alias{ui_import}
\title{User-interface for importing files}
\usage{
ui_import()
}
\value{
An imported data frame formatted and ready to be transformed.
}
\description{
Launches an interactive shiny app for importing data interactively. Also available online at <http://animalnexus.ca> or
by launching the local animalnexus app through \code{animalnexus()}.
}
\examples{
\dontrun{
  my_data <- ui_import()
}

}

