% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trans.R
\name{ui_trans}
\alias{ui_trans}
\title{User-interface for transforming data}
\usage{
ui_trans(r)
}
\arguments{
\item{r}{Data frame. Raw RFID data to transform}
}
\description{
An interactive shiny app for transforming data
}
\details{
See indivdial data transformations for more details: \code{visits()}, \code{presence()}, \code{move()}, \code{disp()}, \code{dom()}, \code{activity()}, and \code{daily()}.
}

