% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{use.layer}
\alias{use.layer}
\title{Add feeding layer to leaflet map}
\usage{
use.layer(map, u, u.scale = 1, u.title = "Time", u.pal = c("yellow",
  "red"), controls = TRUE)
}
\description{
Designed for advanced use (see map.leaflet() for general mapping)
}

