% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{use_layer}
\alias{use_layer}
\title{Add feeding layer to leaflet map}
\usage{
use_layer(map, u, u_scale = 1, u_title = "Time", u_pal = c("yellow",
  "red"), controls = TRUE, u.scale, u.title, u.pal)
}
\description{
Designed for advanced use (see map_leaflet() for general mapping)
}

