% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity}
\alias{activity}
\title{Calculate activity from feeding data}
\usage{
activity(f1, res = 15, by_feeder = FALSE, missing = NULL)
}
\arguments{
\item{f1}{Data frame. A feeding bout data frame for an individual
\code{bird_id}}

\item{res}{Character. The resolution over which to calculate activity. Should
be in the format of "15 min" or "1 hour", etc.}

\item{missing}{Data frame. If there are known times for a particular feeder
for which activity can't be recorded (i.e. times during which a feeder was
inactive).}
}
\description{
This function will take feeding bouts and calculate activity status (active
vs. inactive) at a resolution of \code{res}.
}
\details{
A message will alert you to when the \code{res} is larger than the 50% of the
feeding bouts. This may result in missed feeding bouts, and it may be better
to choose a smaller \code{res}.

The \code{missing} data frame should have columns \code{start} and \code{end}
corresponding to the start and end times of the missing data. Any activity between
those start/end times will be scored as unknown, regardless of the
\code{feeder_id}. However, if \code{by_feeder} is TRUE, \code{missing} may
also include the column \code{feeder_id}. In this case, only activity for the
feeder with the missing start/end times will be scored as unknown. If
\code{by_feeder} is TRUE but \code{missing} does not contain the column
\code{feeder_id}, all activity between the start and end times will be scored
as unknown, regardless of the feeder. See examples.
}

