% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{activity}
\alias{activity}
\title{Activity}
\usage{
activity(
  p,
  res = 15,
  by_logger = FALSE,
  missing = NULL,
  sun = TRUE,
  keep_all = FALSE,
  pass = TRUE,
  f
)
}
\arguments{
\item{p}{Dataframe. A \code{\link[=presence]{presence()}} data frame (may contain multiple
animal_ids).}

\item{res}{Numeric. The resolution over which to calculate activity in
minutes.}

\item{by_logger}{Logical. Should the activity be calculated overall, or
individually for each logger visited? If there is only one logger,
by_logger will automatically revert to TRUE to enable passing of
logger-related variables.}

\item{missing}{Data frame. (NOT AVAILABLE) If there are known times for a
particular logger for which activity can't be recorded (i.e. times during
which a logger was inactive).}

\item{sun}{Logical. Calculate sun rise/set? If by_logger = FALSE, returns
median sun rise/set across all loggers for each day.}

\item{keep_all}{Logical. Keep all individuals, even ones with less than 24hrs of data.}

\item{pass}{Logical. Pass 'extra' columns through the function and append them to the output.}

\item{f}{Depreciated. Use \code{p}.}
}
\description{
Calculate activity status (active vs. inactive) at a resolution of \code{res}
from \code{\link[=presence]{presence()}} data.
}
\details{
A message will alert you to when the \code{res} is larger than the 50\\% of
the presence bout lengths. This may result in missed activity, and it may be
better to choose a smaller \code{res}.

The \code{missing} data frame should have columns \code{start} and \code{end}
corresponding to the start and end times of the missing data. Any activity between
those start/end times will be scored as unknown, regardless of the
\code{logger_id}. However, if \code{by_logger} is TRUE, \code{missing} may
also include the column \code{logger_id}. In this case, only activity for the
logger with the missing start/end times will be scored as unknown. If
\code{by_logger} is TRUE but \code{missing} does not contain the column
\code{logger_id}, all activity between the start and end times will be scored
as unknown, regardless of the logger. See examples.
}
\examples{

v <- visits(chickadees)
p <- presence(v)
a <- activity(p, res = 1)

# By logger (may take a while)
\dontrun{
a <- activity(p, res = 1, by_logger = TRUE)
}

}
