% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chickadees}
\alias{chickadees}
\title{Read data on Chickadees from gap-crossing study}
\format{A data frame with 22095 rows and 6 variables:
\describe{
  \item{experiment}{Which experiment (site) the data was collected at}
  \item{feeder_id}{Unique id of feeder where read was made}
  \item{bird_id}{Unique RFID tag attached to a chickadee}
  \item{time}{Time of read}
  \item{lat}{Latitude of feeder location}
  \item{lon}{Longitude of feeder location}}}
\usage{
chickadees
}
\description{
A dataset containing raw reads of black-capped chickadees visiting feeders in
Prince George, BC, Canada
}
\keyword{datasets}

