% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chickadees}
\alias{chickadees}
\title{Read data on Chickadees from gap-crossing study}
\format{A data frame with 26,753 rows and 10 variables:
\describe{
  \item{animal_id}{Unique RFID tag attached to a chickadee}
  \item{date}{Date of the read}
  \item{time}{Time of read}
  \item{logger_id}{Unique id of logger where read was made}
  \item{experiment}{Which experiment (site) the data was collected at}
  \item{lat}{Latitude of logger location}
  \item{lon}{Longitude of logger location}}}
\usage{
chickadees
}
\description{
A dataset containing raw reads of black-capped chickadees visiting
RFID-enabled feeders (loggers) in Prince George, BC, Canada
}
\keyword{datasets}
