% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_dominance}
\alias{convert_dominance}
\title{Convert displacements for use by functions from the Dominance package}
\usage{
convert_dominance(d)
}
\arguments{
\item{d}{Data frame or List. Either the specific displacements data frame which is
returned as a list item from \code{disp()}, or the whole displacements list
returned by \code{disp()}.}
}
\value{
List of data frames to use in functions from the \link{Dominance}
  package. data_sheet contains all interactions: action.from/action.to
  represent individuals (code is matched to animal_id in the items data
  frame). action.from represent displacers (winners), action.to represent
  displacees (losers), kind.of.action is a dummy variable representing the
  action type (displacement). items is a data frame matching animal_ids to to
  the code used in the data_sheet. actions contain the one action type
  (displacements) and the classification and weighting (corresponds to
  action.from being a winner, and action.to being a loser). bytes is a dummy
  vector indicating that the action type "displacement" should be included in
  the calculation. See examples for specific application.
}
\description{
Converts displacements RFID data into a format for easy use by either the
\code{\link[Dominance]{ADI}} or the \code{\link[asnipe]{Sociogram}} functions
included in the \link{Dominance} package.
}
\examples{
# Calculate displacements
d <- disp(visits(finches_lg))

# Format for use by Dominance package
i <- convert_dominance(d)
i <- convert_dominance(d$displacements) # Equivalent

\dontrun{
# Use Dominance package:
library(Dominance)

# Calculate the Average Dominance Index
ADI(data_sheet = i$data_sheet, items = i$items, actions = i$actions, bytes = i$bytes)

# Construct social network graphs
Sociogram(data_sheet = i$data_sheet, items = i$items, actions = i$actions, bits = i$bytes)
}

}
\seealso{
\link{Dominance} package and it's functions: \link[Dominance]{ADI} and \link[Dominance]{Sociogram}. \url{https://cran.r-project.org/package=Dominance}
}
