% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{disp}
\alias{disp}
\title{Displacements}
\usage{
disp(v, bw = 2, pass = TRUE)
}
\arguments{
\item{v}{Dataframe. A visits data frame containing \strong{all} visits from
\strong{all} animals. From the output of \code{visits}. Must contain columns
\code{animal_id}, \code{logger_id}, \code{start}, and \code{end}.}

\item{bw}{Numeric. The maximum interval in seconds between visits by two
different animals for the interaction to be considered a displacement.}

\item{pass}{Logical. Pass 'extra' columns through the function and append
them to the output.}
}
\value{
A list with the following named items: \enumerate{ \item
  \code{displacements}: A data frame of individual displacement events,
  including the following columns: \itemize{ \item \code{logger_id}: ID of
  the logger at which the event occurred \item ID of the animal being displaced
  (\code{displacee}) \item ID of the animal doing the displacing
  (\code{displacer}) \item Time of the departure of the displacee
  (\code{left}) \item Time of the arrival of the displacer (\code{arrived}) }

  \item \code{summaries}: A data frame of overall wins/lossess per
  individual, containing the following columns: \itemize{ \item ID of the
  animal (\code{animal_id}) \item No. of times the animal was displaced
  (\code{displacee}) \item No. of times the animal was a displacer
  (\code{displacer}) \item Proportion of wins (\code{p_win}) } \item
  \code{interactions}: A data frame of interaction summaries, containing the
  following columns: \itemize{ \item ID of the displacee (\code{displacee})
  \item ID of the displacer (\code{displacer}) \item No. of times this
  interaction occurred (\code{n}) } }
}
\description{
For an entire \code{visits} data frame, identifies displacement events.
Displacements are events when one animal leaves the logger right before the
arrival of another.
}
\details{
The first and last visits on the record are automatically assumed to be
non-displacer and non-displacee, respectively.

In some species displacements can be used to infer dominance. Displacements
The interactions data frame returned by the \code{disp()} function can be
passed directly to the \code{\link[Perc]{as.conflictmat}} function of the
\link{Perc} package to be transformed into a conflict matrix, ready for
analysis of dominance using percolation and conductance. Finally, the
displacements data frame can also be converted using the
\code{\link{convert_anidom}} function to a data frame for use by the
\link{aniDom} package's \link[aniDom]{elo_scores} function.
}
\examples{

# Look at displacements for chickadees in experiment 2
 v <- visits(chickadees[chickadees$experiment == "exp2",])
 d <- disp(v)

 # Look at displacement events:
 d[['displacements']] #or
 d$displacements

 # Look at summaries (identical methods):
 d[['summaries']] #or
 d$summaries

 # Look at interactions (identical methods):
 d[['interactions']] #or
 d$interactions

 # Calculate across different experiments (expect warnings about unequal factor levels):
library(dplyr)

v <- chickadees \%>\%
  group_by(experiment) \%>\%
  do(visits(.))

d <- v \%>\%
  group_by(experiment) \%>\%
  do(data = disp(.))

# Look at the data stored in the 2nd experiment:
d$data[d$experiment == "exp2"][[1]] #or
d[["data"]][[1]] #or
d$data[[1]]

# Access the displacements from the 3rd experiment:
d$data[d$experiment == "exp3"][[1]]$displacements #or
d[["data"]][[2]]$displacements #or
d$data[[2]]$displacements

}
\seealso{
\link{Perc}, \link{aniDom}, \link{Dominance}
}
