% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{disp}
\alias{disp}
\title{'Visits' to 'displacements'}
\usage{
disp(v, bw = 5)
}
\arguments{
\item{v}{Dataframe. A visits data frame containing \strong{all} visits from
\strong{all} birds. From the output of \code{visits}. Must contain columns
\code{bird_id}, \code{feeder_id}, \code{start}, and \code{end}.}

\item{bw}{Numeric. The minimum number of seconds between visits by two
different birds for the event to be considered a displacement.}
}
\value{
A list with the following named items:
\enumerate{
 \item
  \code{displacements}: A data frame of individual displacement events,
  including the following columns:
  \itemize{
   \item ID of the feeder at
  which the event occurred (\code{feeder_id})
   \item ID of the bird being
  displaced (\code{displacee})
   \item ID of the bird doing the displacing
  (\code{displacer})
   \item Time of the departure of the displacee
  (\code{left})
   \item Time of the arrival of the displacer (\code{arrived})
  }

  \item \code{summary}: A data frame of overall wins/lossess per individual,
  containing the following columns:
  \itemize{
   \item ID of the bird (\code{bird_id})
   \item No. of times the bird was displaced (\code{displacee})
   \item No. of times the bird was a displacer (\code{displacer})
   \item Proportion of wins (\code{p_win})
  }
  \item \code{interactions}: A data frame of interaction summaries,
  containing the following columns:
  \itemize{
   \item ID of the displacee (\code{displacee})
   \item ID of the displacer (\code{displacer})
   \item No. of times this interaction occurred (\code{n})
  }
 }
}
\description{
For an entire \code{visits} data frame, identifies displacement events.
Displacements are events when one bird is forced to leave the feeder due to
the imminent arrival of a more dominant bird.
}
\details{
The first and last visits on the record are automatically assumed to be non
displacer and non-displacee, respectively.
}
\examples{
  \dontrun{
    v <- visits(r)
    d <- disp(v)

    # Look at displacement events:
    d[['displacements']][1:5,]
    d$displacements[1:5,]

    # Look at summary:
    d[['summary']]
    d$summary

    # Look at interactions:
    d[['interactions']]
    d$interactions
 }
}

