% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{dom}
\alias{dom}
\title{'displacements' to 'dominance'}
\usage{
dom(d, tries = 50, omit_zero = TRUE)
}
\arguments{
\item{d}{Data frame or List. Either the interactions data frame which is
returned as a list item from \code{disp()}, or the whole displacements list
returned by \code{disp()}.}

\item{tries}{Numeric. The maximum number of iterations to find the 'best guess'}

\item{omit_zero}{Logical. Should individuals with 0 interactions (sum of wins
and losses) be omitted?}
}
\description{
'displacements' to 'dominance'
}

