% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{feeding}
\alias{feeding}
\title{Presence at a logger}
\usage{
feeding(v, bw = 15, pass = TRUE)
}
\arguments{
\item{v}{Dataframe. A visits data frame from the output of \code{visits} (may
contain more than one bird_id). Must contain columns \code{bird_id},
\code{feeder_id}, \code{start}, and \code{end}.}

\item{bw}{Numeric. The minimum number of minutes between visits for two
successive visits to be considered separate feeding bouts. When \code{bw} =
NULL only visits to another feeder are scored as a separate feeding bout.}

\item{pass}{Logical. Pass 'extra' columns through the function and append
them to the output.}
}
\value{
A data frame of feeding bouts. A feeding bout is defined as the bout
  of time spent making regular visits to a feeder, in which each visit
  occurrred within some cutoff time of the last. This data frame has the
  following columns:
  \itemize{
  \item ID of the bird (\code{bird_id})
  \item ID of the feeder(\code{feeder_id})
  \item Time of the start of the feeding bout (\code{feed_start})
  \item Time of the end of the feeding bout (\code{feed_end})
  }
}
\description{
Turns visits at mulitple feeders into feeding bouts. Feeding bouts are
separated by switching feeders (when \code{bw = NULL}). If \code{bw} is not
\code{NULL}, they are separated by switching feeders or by \code{bw} minutes.
}
\examples{

v <- visits(finches)
f <- feeding(v)

# Summarize a movement dataframe (get total time spent feeding)
library(dplyr)
f.totals <- f \%>\%
    group_by(bird_id, feeder_id) \%>\%
    summarize(feed_length = sum(feed_length))

# Calculate across different experiments (expect warnings about unequal factor levels):
library(dplyr)

v <- chickadees \%>\%
  group_by(experiment) \%>\%
  do(visits(.))

f <- v \%>\%
  group_by(experiment) \%>\%
  do(feeding(.))


}

