% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedr-package.R
\docType{package}
\name{feedr-package}
\alias{feedr-package}
\alias{feedr}
\title{Transforming Raw RFID Data}
\description{
Load, clean, transform, and visualize raw RFID data for looking at a variety
of biological questions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/animalnexus/feedr}
  \item \url{https://animalnexus.github.io/feedr/}
}

}
\author{
\strong{Maintainer}: Steffi LaZerte \email{steffi@steffi.ca}

}
\keyword{internal}
