% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedr.R
\docType{package}
\name{feedr-package}
\alias{animalnexus-package}
\alias{feedr}
\alias{feedr-package}
\title{Transforming Raw RFID Data}
\description{
Load, clean, transform, and visualize raw RFID data for looking at a variety
of biological questions.
}
\references{
animalnexus \url{http://animalnexus.ca}.
}

