% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedr-package.R
\docType{package}
\name{feedr-package}
\alias{feedr}
\alias{feedr-package}
\title{feedr: Livestock Diet Formulation and Optimization}
\description{
Open-source least-cost diet formulation for livestock nutritionists and researchers. Pipe-first API with a DuckDB backend.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/austin-putz/feedr}
  \item Report bugs at \url{https://github.com/austin-putz/feedr/issues}
}

}
\author{
\strong{Maintainer}: Austin Putz \email{putz.austin@gmail.com}

Authors:
\itemize{
  \item Austin Putz \email{putz.austin@gmail.com}
}

}
\keyword{internal}
