% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.web}
\alias{load.web}
\title{Load read data from the web}
\usage{
load.web(r_file, tz = "Canada/Pacific", sep = ",")
}
\arguments{
\item{r_file}{Character. The location of a single file to load.}

\item{tz}{Character. The time zone the date/times should be converted to (should match one of the zones produced by \code{OlsonNames())}.}

\item{sep}{Character. An override for the separator in the \code{read.table()} call (see \code{sep =} under \code{?read.table} for more details).}
}
\description{
Loads raw read data from csv files downloaded from the \url{http://gaia.tru.ca/birdMOVES/datadownload.html} website and formats them for use with the feedr functions. This is merely a wrapper function that does many things that you can do yourself. It's utility depends on how standardized your data is, and whether you have extra details you need to address.
}
\examples{
\dontrun{r <- load.web("downloaded_file.csv")}
}

