% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_format}
\alias{load_format}
\title{Format data}
\usage{
load_format(r, tz = Sys.timezone(), tz_disp = NULL, dst = FALSE,
  time_format = "ymd HMS")
}
\arguments{
\item{r}{Data frame. Data frame to format.}

\item{tz}{Character. The time zone the date/times are in (should match one of
the zones produced by \code{OlsonNames())}. Attempts to use user's system
timezone, if none supplied. Defaults to UTC if all else fails.}

\item{tz_disp}{Character. The time zone the date/times should be displayed in
(if not the same as \code{tz}; should match one of the zones produced by
\code{OlsonNames())}.}

\item{dst}{Logical. Whether or not to use Daylight Savings. When set to FALSE
timezones are converted to the Etc/GMT+X timezones which do not include
DST. (Note this overrides the timezone specification such that a timezone
of America/Vancouver, which would normally include DST in the summer, will
be transformed to a timezone with the same GMT offset, but not including
DST).}

\item{time_format}{Character. The time format of the 'time' column. Defaults
to "ymd HMS". Should be in formats usable by the \code{parse_date_time()}
function from the lubridate package (e.g., "ymd HMS", "mdy HMS", "dmy HMS",
etc.).}
}
\description{
Formats manually loaded data. Not necessary if using any of the helper loading functions (e.g., \code{load_raw()}, \code{load_raw_all()}, or \code{data_dl()}.
}

