% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map.ggmap}
\alias{map.ggmap}
\title{Map data using ggmap}
\usage{
map.ggmap(u = NULL, p = NULL, locs = NULL, u.scale = 1, p.scale = 1,
  u.title = "Time", p.title = "Path use", u.pal = c("yellow", "red"),
  p.pal = c("yellow", "red"), maptype = "satellite", mapsource = "google",
  zoom = 17, which = NULL)
}
\arguments{
\item{u}{Dataframe. Summarized feeding data with columns \code{feeder_id} and
\code{amount}. It can also contain \code{bird_id} for individual-based
data, and lat/lon instead of a locs argument.}

\item{p}{Dataframe. Summarized movement data with columns \code{feeder_id},
\code{move_path}, and \code{path_use}. It can also contain \code{bird_id}
for individual-based data, and lat/lon instead of a locs argument.}

\item{locs}{Dataframe. Lat and long for each feeder_id, required if lat and
lon not in either u or p.}

\item{u.scale, p.scale}{Numerical. Scaling constants to increase (> 1) or
decrease (< 1) the relative size of use (u) and path (p) data.}

\item{u.title, p.title}{Character. Titles for the legends of use (u) and
path (p) data.}

\item{u.pal, p.pal}{Character vectors. Colours used to construct gradients for
use (u) and movement (p) data.}

\item{maptype}{Character. The type of map to download. See \code{maptype}
under \code{\link[ggmap]{get_map}} for more options.}

\item{mapsource}{Character. The source of the map to download. See
\code{source} under \code{\link[ggmap]{get_map}} for more options.}

\item{zoom}{Numeric. The zoom level of the map to download. See \code{zoom}
under \code{\link[ggmap]{get_map}} for more options.}

\item{which}{Character vector. A vector of bird ids specifying which to show.
Only applies when using individual data.}
}
\value{
An interactive leaflet map with layers for feeding time, movement
  paths and feeder positions.
}
\description{
Visualize feeder use and path data using ggmap in R. This produces a
static map. The user must summarize feeder use and path data in the manner
that they wish to visualize it. This function can take invidiual bird data as
well as grand summarise (see Details and Examples).
}
\examples{
\dontrun{

v <- visits(finches)
f <- ddply(v, .(bird_id), feeding, bw = 15)
m <- ddply(v, .(bird_id), move)

# Summarise data for visualization (use totals):
f.all <- ddply(f, .(feeder_id), summarise,
           amount = sum(feed_length) / bird_n)

m.all <- ddply(m, .(move_path), summarise,
           path_use = length(move_path) / bird_n)

# Look at total summary maps
map.leaflet(u = f.all, p = m.all)
map.ggmap(u = f.all, p = m.all)

# Summarise data for visualization (use individuals):
f.indiv <- ddply(f, .(bird_id, feeder_id), summarise,
           amount = sum(feed_length))

m.indiv <- ddply(m, .(bird_id, move_path), summarise,
           path_use = length(move_path))

# Look at individual summary maps (note that Leaflet just stacks individuals
one on top of the other)
map.leaflet(u = f.indiv, p = m.indiv)
map.ggmap(u = f.indiv, p = m.indiv, u.scale = 0.7, p.scale = 0.05)
}
}

