% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{move}
\alias{move}
\title{Movements}
\usage{
move(v, all = FALSE, pass = TRUE)
}
\arguments{
\item{v}{Dataframe. A visits data frame (may contain multiple animal_ids). From
the output of \code{visits}. Must contain columns \code{animal_id},
\code{logger_id}, \code{start}, and \code{end}.}

\item{all}{Logical. Should all animal_ids be returned, even if the animal made no
movements? If TRUE, a data frame with NAs for all columns except
\code{animal_id} will be returned, if FALSE, an empty data frame will be
returned.}

\item{pass}{Logical. Pass 'extra' columns through the function and append
them to the output.}
}
\value{
A data frame of movements. These are defined as the bout of time from
  leaving one logger to arriving at a second one. Each movement bout consists
  of two rows of data containing:
  \itemize{
  \item ID of the animal (\code{animal_id})
  \item Date of event (\code{date})
  \item Time of event (\code{time})
  \item The ID of loggers involved (\code{logger_id})
  \item The movement path including direction (\code{move_dir})
  \item The movement path independent of direction (\code{move_path})
  \item The 'strength' of the connection (inverse of time taken to move
  between; \code{strength})
  \item Information on whether left/arrived (\code{direction})
  \item The ID of a single move event for a particular individual (\code{move_id})
  \item Any extra columns \code{pass}ed through
  }
}
\description{
Turns visits to mulitple loggers into movements between loggers
}
\examples{

v <- visits(finches)
m <- move(v)

# Summarize (divide by 2 because 2 rows per event)
library(dplyr)

m.totals <- m \%>\%
  group_by(animal_id, move_path) \%>\%
  summarize(n_path = length(move_path) /2)

# Calculate across different experiments:

v <- chickadees \%>\%
  group_by(experiment) \%>\%
  do(visits(.))

m <- v \%>\%
  group_by(experiment) \%>\%
  do(move(.))

}
