% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{move}
\alias{move}
\title{'Visits' to 'movements'}
\usage{
move(v, all = FALSE, pass = TRUE)
}
\arguments{
\item{v}{Dataframe. A visits data frame (may contain multiple bird_ids). From
the output of \code{visits}. Must contain columns \code{bird_id},
\code{feeder_id}, \code{start}, and \code{end}.}

\item{all}{Logical. Should all bird_ids be returned, even if the bird made no
movements? If TRUE, a data frame with NAs for all columns except
\code{bird_id} will be returned, if FALSE, an empty data frame will be
returned.}

\item{pass}{Logical. Pass 'extra' columns through the function and append
them to the output.}
}
\value{
A data frame of movements. These are defined as the bout of time from
  leaving one feeder to arriving at a second one. Each movement bout consists
  of two rows of data containing:
  \itemize{
  \item ID of the bird (\code{bird_id})
  \item Time of event (\code{time})
  \item The ID of feeders involved (\code{feeder_id})
  \item The movement path including direction (\code{move_dir})
  \item The movement path independent of direction (\code{move_path})
  \item The 'strength' of the connection (inverse of time taken to move
  between; \code{strength})
  \item Information on whether left/arrived (\code{direction})
  \item The ID of a single move event for a particular individual (\code{move_id})
  \item Any extra columns \code{pass}ed through
  }
}
\description{
Turns visits to mulitple feeders into movements between feeders
}
\examples{

v <- visits(finches)
m <- move(v)

# Summarize (divide by 2 because 2 rows per event)
library(dplyr)

m.totals <- m \%>\%
  group_by(bird_id, move_path) \%>\%
  summarize(n_path = length(move_path) /2)
}

