% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{path_layer}
\alias{path_layer}
\title{Add movement layer to leaflet map}
\usage{
path_layer(map, m, m_scale = 1, m_title = "Path use", m_pal = c("yellow",
  "orange", "red"), controls = TRUE, p.scale, p.title, p.pal)
}
\description{
Designed for advanced use (see map_leaflet() for general mapping)
}

