% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{visits}
\alias{visits}
\title{'Raw' data to 'visits' data}
\usage{
visits(r, bw = 3, allow.imp = FALSE, na.rm = FALSE)
}
\arguments{
\item{r}{Dataframe. Contains raw reads from an RFID reader with colums \code{bird_id}, \code{feeder_id}, \code{time}.}

\item{bw}{Numerical. The minimum number of seconds between reads for two successive reads to be considered separate visits.}

\item{allow.imp}{Logical. Whether impossible visits should be allowed (see details).}

\item{na.rm}{Logical. Whether NA values should be automatically omited. Otherwise an error is returned.}
}
\value{
A data frame with visits specifying \code{bird_id} and \code{feeder_id} as well as the \code{start} and \code{end} of the visit. Any extra columns that are unique at the level of bird_id or feeder_id will also be passed through (i.e. age, sex, feeder location, etc.).
}
\description{
Raw data from RFID feeders contain multiple reads per individual simply
because the individual sat there long enough. In \code{visits()} these reads
are collapsed into one visit.
}
\details{
Visits are defined by three pieces of data:
\itemize{
\item How much time has passed between reads (\code{bw})
\item A change in identity between two successive reads (\code{bird_id})
\item A change in feeder for a single individual (\code{feeder_id})
}

The function will return an error if impossible visits are detected (unless \code{allow.imp = TRUE}) . A visit is deemed impossible if a single bird travels between feeders in less time than specified by \code{bw}.
}
\examples{
\dontrun{
r <- load.web("downloaded_file.csv")
v <- visits(r)
}
}

