% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{visits}
\alias{visits}
\title{Visits}
\usage{
visits(r, bw = 3, allow_imp = FALSE, bw_imp = 2, na_rm = FALSE,
  pass = TRUE, allow.imp, na.rm)
}
\arguments{
\item{r}{Dataframe. Contains raw reads from an RFID reader with colums
\code{animal_id}, \code{logger_id}, \code{time}.}

\item{bw}{Numerical. The minimum interval, in seconds, between reads for two
successive reads to be considered separate visits.}

\item{allow_imp}{Logical. Whether impossible visits should be allowed (see
details).}

\item{bw_imp}{Numerical. The minimum number of seconds required to travel
between loggers. If quicker, visits considered impossible.}

\item{na_rm}{Logical. Whether NA values should be automatically omited.
Otherwise an error is returned.}

\item{pass}{Logical. Pass 'extra' columns through the function and append
them to the output.}

\item{allow.imp, na.rm}{Depreciated.}
}
\value{
A data frame with visits specifying \code{animal_id} and
  \code{logger_id} as well as the \code{start} and \code{end} of the visit.
  Any extra columns that are unique at the level of animal_id or logger_id will
  also be passed through (i.e. age, sex, logger location, etc.).
}
\description{
Raw data from RFID loggers contain multiple reads per individual simply
because the individual sat there long enough. In \code{visits()} these reads
are collapsed into one visit.
}
\details{
Visits are defined by three pieces of data:
\itemize{
\item How much time has passed between reads (\code{bw})
\item A change in identity between two successive reads (\code{animal_id})
\item A change in logger for a single individual (\code{logger_id})
}

The function will return an error if impossible visits are detected (unless
  \code{allow_imp = TRUE}) . A visit is deemed impossible if a single animal
  travels between loggers in less time than specified by \code{bw}.
}
\examples{
v <- visits(finches)
head(v)

v <- visits(finches, bw = 30)
head(v)

# Calculate across different experiments:
library(dplyr)
v <- chickadees \%>\%
  group_by(experiment) \%>\%
  do(visits(.))
}
